IE ration control
With only one limit switch. Requires that you manually move the piston to a position just before triggering the limit switch on the inhalation cycle in order for the device to assign the correct duration to inhalation/exhalation.

Pressure sensors also activate, but not being polled or stored.

Maximum output pressure is around +45mbar  on exhale and -30mbar on inhale